Fixed minor errors - Changed type of loop iterator to resolve comparison warning - Damageable no longer inherits from Actor to resolve ambiguity errors

+++ src/amor.cpp
@@ -51,7 +51,7 @@ void Armor::init(std::string ID)
   Json::Value items = value["armors"];
 
   //go through the json file and find the item by ID
-  for (int i = 0; i < items.size(); i++)
+  for (size_t i = 0; i < items.size(); i++)
   {
     if (ID.compare(items[i]["id"].asCString()) == 0)
     {
